agent/requirements.txt
aiortc>=1.6.0
mss>=9.0.0
pynput>=1.7.6
websockets>=12.0
aiohttp>=3.9.0
av>=11.0.0
Pillow>=10.0.0
numpy>=1.24.0
